requirements.txt
tensorflow
tensorflow_probability
matplotlib
absl-py
scipy
seaborn
sys
pandas
os
scipy
time

